patch 8.0.1428: compiler warning on 64 bit MS-Windows system
Problem:    Compiler warning on 64 bit MS-Windows system.
Solution:   Change type from "int" to "size_t". (Mike Williams)

+++ src/ex_getln.c
@@ -180,7 +180,7 @@ abandon_cmdline(void)
     static int
 empty_pattern(char_u *p)
 {
-    int n = STRLEN(p);
+    size_t n = STRLEN(p);
 
     /* remove trailing \v and the like */
     while (n >= 2 && p[n - 2] == '\\'
